Add arm64 support

+++ build.sh
@@ -2,28 +2,46 @@
 
 set -e
 
+if [[ "$1" != "clean" && "$1" != "amd64" && "$1" != "arm64" ]]; then
+	echo "Usage: ./build.sh [amd64/arm64] or ./build.sh clean"
+	exit 0
+fi
+
+if [ "$1" = "amd64" ]; then
+	URL=https://cdimage.debian.org/debian-cd/current/amd64/iso-cd/debian-13.5.0-amd64-netinst.iso
+	SHASUM="b2be60c555e328b4fa5ebb2d0e5c7ee6bc3eb4250c4dcfd3f78b8d9aec596efdf9f14f10a898c280eb252d50bbac91ea0a2bba29736df0d4985d50d4c8d77519  debian-13.5.0-amd64-netinst.iso"
+	ISOISTDIR=amd
+fi
+
+if [ "$1" = "arm64" ]; then
+	URL=https://cdimage.debian.org/debian-cd/current/arm64/iso-cd/debian-13.5.0-arm64-netinst.iso
+	SHASUM="e81aa710007e5d6cf05da300431223a3f75ed7264f244cb374f59c50037f1d96056e378894768c710e6f165c621f3ad7ad0fbc0cc01084488d797788237d8b2b  debian-13.5.0-arm64-netinst.iso"
+	ISOISTDIR=a64
+fi
+
 if [ "$1" = "clean" ]; then
    sudo rm -rf build
-   rm -f DATUM-box.iso
+   rm -f DATUM-box-*.iso
    exit 0
 fi
 
 mkdir build
 cd build
 
-wget https://cdimage.debian.org/debian-cd/current/amd64/iso-cd/debian-13.5.0-amd64-netinst.iso
-echo "b2be60c555e328b4fa5ebb2d0e5c7ee6bc3eb4250c4dcfd3f78b8d9aec596efdf9f14f10a898c280eb252d50bbac91ea0a2bba29736df0d4985d50d4c8d77519  debian-13.5.0-amd64-netinst.iso" | sha512sum -c
+wget $URL
+echo $SHASUM | sha512sum -c
 
-xorriso -osirrox on -indev debian-13.5.0-amd64-netinst.iso -extract / debianfiles
-chmod +w -R debianfiles/install.amd/
-gunzip debianfiles/install.amd/initrd.gz
+xorriso -osirrox on -indev debian-13.5.0-$1-netinst.iso -extract / debianfiles
+chmod +w -R debianfiles/install.$ISOISTDIR/
+gunzip debianfiles/install.$ISOISTDIR/initrd.gz
 cp ../preseed.cfg ../post_config.sh ../happen_bashrc .
-echo preseed.cfg | cpio -H newc -o -A -F debianfiles/install.amd/initrd
-echo post_config.sh | cpio -H newc -o -A -F debianfiles/install.amd/initrd
-echo happen_bashrc | cpio -H newc -o -A -F debianfiles/install.amd/initrd
-gzip debianfiles/install.amd/initrd
-chmod -w -R debianfiles/install.amd/
+echo preseed.cfg | cpio -H newc -o -A -F debianfiles/install.$ISOISTDIR/initrd
+echo post_config.sh | cpio -H newc -o -A -F debianfiles/install.$ISOISTDIR/initrd
+echo happen_bashrc | cpio -H newc -o -A -F debianfiles/install.$ISOISTDIR/initrd
+gzip debianfiles/install.$ISOISTDIR/initrd
+chmod -w -R debianfiles/install.$ISOISTDIR/
 
+if [ "$1" = "amd64" ]; then
 chmod +w debianfiles/isolinux/isolinux.cfg
 cat > debianfiles/isolinux/isolinux.cfg << 'EOF'
 default auto
@@ -36,15 +54,16 @@ label auto
   append initrd=/install.amd/initrd.gz auto=true priority=critical --- quiet
 EOF
 chmod -w debianfiles/isolinux/isolinux.cfg
+fi
 
 chmod +w debianfiles/boot/grub/grub.cfg
-cat > debianfiles/boot/grub/grub.cfg << 'EOF'
+cat > debianfiles/boot/grub/grub.cfg << EOF
 set default=0
 set timeout=1
 
 menuentry "Automated Install" {
-    linux  /install.amd/vmlinuz auto=true priority=critical --- quiet
-    initrd /install.amd/initrd.gz
+    linux  /install.$ISOISTDIR/vmlinuz auto=true priority=critical --- quiet
+    initrd /install.$ISOISTDIR/initrd.gz
 }
 EOF
 chmod -w debianfiles/boot/grub/grub.cfg
@@ -55,11 +74,12 @@ find -follow -type f ! -name md5sum.txt -print0 | xargs -0 md5sum > md5sum.txt
 chmod -w md5sum.txt
 cd ..
 
-dd if=debian-13.5.0-amd64-netinst.iso bs=1 count=432 of=isohdpfx.bin
-
-xorriso -as mkisofs \
+if [ "$1" = "amd64" ]; then
+	dd if=debian-13.5.0-amd64-netinst.iso bs=1 count=432 of=isohdpfx.bin
+	
+	xorriso -as mkisofs \
 		-r -V 'Debian 13.5.0 amd64 n' \
-		-o ../DATUM-box.iso \
+		-o ../DATUM-box-amd64.iso \
 		-J -joliet-long -cache-inodes \
 		-isohybrid-mbr isohdpfx.bin \
 		-b isolinux/isolinux.bin -c isolinux/boot.cat \
@@ -68,3 +88,18 @@ xorriso -as mkisofs \
 		-e boot/grub/efi.img \
 		-no-emul-boot -isohybrid-gpt-basdat -isohybrid-apm-hfsplus \
 		debianfiles
+fi
+
+if [ "$1" = "arm64" ]; then
+	dd if=debian-13.5.0-arm64-netinst.iso bs=512 skip=1427456 count=8192 of=efi.img
+	
+	xorriso -as mkisofs \
+		-r -V 'Debian 13.5.0 arm64 n' \
+		-o ../DATUM-box-arm64.iso \
+		-J -joliet-long -cache-inodes \
+		-e boot/grub/efi.img \
+		-no-emul-boot \
+		-append_partition 2 0xef efi.img \
+		-partition_cyl_align all \
+		debianfiles
+fi
